feat: merge upgraders for created and updated system fields
